requirements.txt
logzero==1.7.0
pandas==1.5.3
protobuf==4.23.4
python-dotenv==1.0.1
streamlit==1.31.1
Requests==2.31.0